metrics: MetricsReporter started
Fixes #5

WIP to report metrics to StackDriver monitoring
about the inflight HTTP requests.

The main abstraction is MetricsReporter with methods
- NewMetricsReporter(ctx context.Context, period time.Duration)
- SetStatsReceiver(func(stats *sproxy.InflightStats) error
- AddEvent(sproxy.Event)
- go Do()
- Close()

which should be able to initiate a metrics reporter
that receives HTTP event information, aggregates information
periodically and can be stopped at will with .Close()

Here is a sample to show how it works with a simulation:
```go
package main

import (
	"encoding/json"
	"log"
	"math/rand"
	"time"

	sproxy "github.com/GoogleCloudPlatform/stackdriver-reverse-proxy"
)

func main() {
	mr, err := sproxy.NewMetricsReporter(550*time.Millisecond)
	if err != nil {
		log.Fatalf("metricsReporter: %v", err)
	}
	mr.SetStatsReceiver(func(stats *sproxy.InflightStats) error {
		blob, _ := json.MarshalIndent(stats, "", "  ")
		log.Printf("statsIn: %s", blob)
		return nil
	})
	go mr.Do()
	defer mr.Close()

	n := 100000
	doneChan := make(chan bool, n)
	for i := 0; i < n; i++ {
		go func(ii int) {
			defer func() {
				doneChan <- true
			}()
			startTime := time.Now()
			evt := &sproxy.Event{
				StartTime:  startTime,
				StatusCode: rand.Intn(600) + 100,
			}

			evt.Request = rand.Intn(37)%2 == 0
			evt.EndTime = time.Now()
			mr.AddEvent(evt)
		}(i)

		refreshTime := time.Duration((1 + rand.Intn(500))) * time.Millisecond
		<-time.After(refreshTime)
	}

	for i := 0; i < n; i++ {
		<-doneChan
	}
}
```

What is left is just doing the raw StackDriver monitoring client
initialization and organizing the metrics from sproxy.InflightStats
and sending them along.